ml-vim/requirements.txt
# Function dependencies
Pillow >= 5.2.0
google-cloud-storage
google-cloud-vision